% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{expand_grid_df}
\alias{expand_grid_df}
\title{Get all combinations of tibbles.}
\usage{
expand_grid_df(...)
}
\arguments{
\item{...}{a bunch of tibbles}
}
\value{
A tibble which represents all combinations of the tibbles given
}
\description{
Get all combinations of tibbles.
}
\examples{
expand_grid_df(tibble(a = 1:3, b = 2:4), tibble(c = 5:3))
}
